requirements.txt
flask==1.0.2
mysql-connector==2.1.6
flask-cors==3.0.4
